refactor: redirect all APIMethods* references to their Http4s equivalents
Every APIMethodsNNN.scala left in the tree is now a pure runtime shim
-- a literal alias (val ImplementationsX = Http4sNNN.ImplementationsX)
plus an empty trait, kept only so existing imports still compile. This
redirects every remaining production and test reference to the real
Http4sNNN object directly, clearing the way to delete the 12 shim
files entirely in a follow-up commit.

Production code:
- ResourceDocsAPIMethods.scala: dropped the 6 empty-trait mixins
  (APIMethods220/210/200/140/130/121, confirmed empty by direct read)
  and its 9 APIMethods400.Implementations4_0_0.X call sites, now
  Http4s400.Implementations4_0_0.X.
- APIUtil.scala: two nameOf(...) fully-qualified references
  (APIMethods510/500 -> Http4s510/500).
- OBPAPI5_1_0.scala / OBPAPI6_0_0.scala / OBPAPI5_0_0.scala: dropped
  the APIMethods310/400/500/510 imports and mixins (all confirmed
  empty), consolidated the now-orphaned "re-export so this compiles
  after APIMethods* was replaced with a stub" comments into one
  accurate note per file (the stub isn't just replaced now, it's gone).

~41 test files: mechanical import + reference swap
(code.api.vX.APIMethodsNNN -> code.api.vX.Http4sNNN;
APIMethodsNNN.ImplementationsX -> Http4sNNN.ImplementationsX) across
v3_1_0/v4_0_0/v5_0_0/v5_1_0/v6_0_0 test suites. These are literal
aliases to the same object, so this is a pure text substitution with
no behavior change -- verified by a full test-compile afterward.

Cleaned up stale comments/docs that would otherwise reference deleted
files after the follow-up deletion commit: Http4s600.scala,
Http4s210.scala, Glossary.scala (incl. an ASCII diagram), RateLimitingUtil.scala,
ChatEventPublisher.scala, FAQ.md, and introductory_system_documentation.md
(including removing a logback logger entry targeting a class that
would no longer exist). Left the "Intentional drift from Lift's
APIMethodsNNN.scala source-of-truth" comments elsewhere untouched --
those are historical citations of where a documented behavioral
decision came from, not claims that the file currently exists, so
they stay accurate after the deletion.
