+++ obp-api/src/test/scala/code/api/v4_0_0/ApiCollectionEndpointTest.scala
@@ -29,7 +29,7 @@ import org.json4s._
 import code.api.ResourceDocs1_4_0.SwaggerDefinitionsJSON
 import code.api.berlin.group.v1_3.Http4sBGv13Alias
 import code.api.util.APIUtil.OAuth._
-import code.api.v4_0_0.APIMethods400.Implementations4_0_0
+import code.api.v4_0_0.Http4s400.Implementations4_0_0
 import com.github.dwickern.macros.NameOf.nameOf
 import com.openbankproject.commons.util.ApiVersion
 import org.json4s.native.Serialization.write
