+++ obp-api/src/test/scala/code/api/v4_0_0/ApiCollectionTest.scala
@@ -30,7 +30,7 @@ import code.api.ResourceDocs1_4_0.SwaggerDefinitionsJSON
 import code.api.util.APIUtil.OAuth._
 import code.api.util.ApiRole
 import code.api.util.ErrorMessages.{ApiCollectionEndpointNotFound, UserHasMissingRoles, AuthenticatedUserIsRequired}
-import code.api.v4_0_0.APIMethods400.Implementations4_0_0
+import code.api.v4_0_0.Http4s400.Implementations4_0_0
 import code.entitlement.Entitlement
 import com.github.dwickern.macros.NameOf.nameOf
 import com.openbankproject.commons.util.ApiVersion
