+++ obp-api/src/test/scala/code/api/v4_0_0/BankTests.scala
@@ -8,7 +8,7 @@ import code.api.util.APIUtil.OAuth._
 import code.api.util.ApiRole.CanCreateBank
 import code.api.util.ErrorMessages.UserHasMissingRoles
 import code.api.util.{ApiRole, ErrorMessages, NewStyle}
-import code.api.v4_0_0.APIMethods400.Implementations4_0_0
+import code.api.v4_0_0.Http4s400.Implementations4_0_0
 import code.entitlement.Entitlement
 import code.setup.{APIResponse, DefaultUsers}
 import com.github.dwickern.macros.NameOf.nameOf
