+++ obp-api/src/test/scala/code/api/v4_0_0/ConnectorMethodTest.scala
@@ -31,7 +31,7 @@ import code.api.util.ApiRole._
 import code.api.util.APIUtil.OAuth._
 import code.api.util.ErrorMessages.{ConnectorMethodAlreadyExists, UserHasMissingRoles}
 import code.api.util.{ApiRole, CallContext}
-import code.api.v4_0_0.APIMethods400.Implementations4_0_0
+import code.api.v4_0_0.Http4s400.Implementations4_0_0
 import code.bankconnectors.InternalConnector
 import code.connectormethod.{ConnectorMethodProvider, JsonConnectorMethod}
 import code.entitlement.Entitlement
