+++ obp-api/src/test/scala/code/api/v4_0_0/EntitlementTests.scala
@@ -11,7 +11,7 @@ import code.setup.DefaultUsers
 import code.api.util.APIUtil.OAuth._
 import code.api.util.ExampleValue.bankIdExample
 import code.api.v2_0_0.CreateEntitlementJSON
-import code.api.v4_0_0.APIMethods400.Implementations4_0_0
+import code.api.v4_0_0.Http4s400.Implementations4_0_0
 import com.github.dwickern.macros.NameOf.nameOf
 import com.openbankproject.commons.util.ApiVersion
 import org.json4s.native.Serialization.write
