+++ obp-api/src/test/scala/code/api/v4_0_0/GetScannedApiVersionsTest.scala
@@ -27,7 +27,7 @@ package code.api.v4_0_0
 
 import code.api.util.APIUtil
 import code.api.util.ApiRole._
-import code.api.v4_0_0.APIMethods400.Implementations4_0_0
+import code.api.v4_0_0.Http4s400.Implementations4_0_0
 import code.entitlement.Entitlement
 import code.setup.PropsReset
 import com.github.dwickern.macros.NameOf.nameOf
