+++ obp-api/src/test/scala/code/api/v4_0_0/PasswordRecoverTest.scala
@@ -33,7 +33,7 @@ import code.api.util.APIUtil.OAuth._
 import code.api.util.ApiRole._
 import com.openbankproject.commons.util.ApiVersion
 import code.api.util.ErrorMessages._
-import code.api.v4_0_0.APIMethods400.Implementations4_0_0
+import code.api.v4_0_0.Http4s400.Implementations4_0_0
 import code.entitlement.Entitlement
 import code.model.dataAccess.{AuthUser, ResourceUser}
 import code.users.Users
