+++ obp-api/src/test/scala/code/api/v4_0_0/TransactionRequestsTest.scala
@@ -16,7 +16,7 @@ import code.api.util.{APIUtil, ErrorMessages}
 import code.api.v1_4_0.JSONFactory1_4_0.{ChallengeAnswerJSON, TransactionRequestAccountJsonV140}
 import code.api.v2_0_0.TransactionRequestBodyJsonV200
 import code.api.v2_1_0._
-import code.api.v4_0_0.APIMethods400.Implementations4_0_0
+import code.api.v4_0_0.Http4s400.Implementations4_0_0
 import code.api.v5_0_0.PhysicalCardJsonV500
 import code.api.v5_1_0.{AgentJsonV510, PutAgentJsonV510}
 import code.bankconnectors.Connector
