+++ obp-api/src/test/scala/code/api/v5_1_0/AtmAttributeTest.scala
@@ -6,7 +6,7 @@ import code.api.util.APIUtil.OAuth._
 import code.api.util.ApiRole._
 import code.api.util.{ApiRole, ErrorMessages}
 import code.api.util.ErrorMessages.{AtmNotFoundByAtmId, UserHasMissingRoles}
-import code.api.v5_1_0.APIMethods510.Implementations5_1_0
+import code.api.v5_1_0.Http4s510.Implementations5_1_0
 import code.entitlement.Entitlement
 import code.setup.DefaultUsers
 import com.github.dwickern.macros.NameOf.nameOf
