+++ obp-api/src/test/scala/code/api/v5_1_0/ConsentObpTest.scala
@@ -32,7 +32,7 @@ import code.api.util.ApiRole._
 import code.api.util.ErrorMessages._
 import code.api.util.{APIUtil, Consent}
 import code.api.util.APIUtil.OAuth._
-import code.api.v3_0_0.{APIMethods300, UserJsonV300}
+import code.api.v3_0_0.{Http4s300, UserJsonV300}
 import code.api.v3_1_0.{ConsentJsonV310, PostConsentChallengeJsonV310, PostConsentEntitlementJsonV310, PostConsentViewJsonV310}
 import code.api.v3_1_0.OBPAPI3_1_0.Implementations3_1_0
 import code.api.v5_1_0.OBPAPI5_1_0.Implementations5_1_0
@@ -57,7 +57,7 @@ class ConsentObpTest extends V510ServerSetup {
   object AnswerConsentChallenge extends Tag(nameOf(Implementations3_1_0.answerConsentChallenge))
 
   object VersionOfApi2 extends Tag(ApiVersion.v3_0_0.toString)
-  object GetUserByUserId extends Tag(nameOf(APIMethods300.Implementations3_0_0.getUserByUserId))
+  object GetUserByUserId extends Tag(nameOf(Http4s300.Implementations3_0_0.getUserByUserId))
 
   val validHeaderConsumerKey = List((RequestHeader.`Consumer-Key`, user1.map(_._1.key).getOrElse("SHOULD_NOT_HAPPEN")))
