+++ obp-api/src/test/scala/code/api/v6_0_0/CustomViewsTest.scala
@@ -5,7 +5,7 @@ import code.api.util.APIUtil.OAuth._
 import code.api.util.ApiRole.{CanCreateCustomView, CanGetCustomViews}
 import code.api.util.ErrorMessages
 import code.api.util.ErrorMessages.{InvalidCustomViewFormat, InvalidJsonFormat, UserHasMissingRoles}
-import code.api.v6_0_0.APIMethods600.Implementations6_0_0
+import code.api.v6_0_0.Http4s600.Implementations6_0_0
 import code.entitlement.Entitlement
 import code.setup.DefaultUsers
 import code.views.system.ViewDefinition
