+++ obp-api/src/test/scala/code/api/v6_0_0/GroupEntitlementsTest.scala
@@ -8,7 +8,7 @@ import code.api.util.ApiRole.{
 }
 import code.api.util.ErrorMessages
 import code.api.util.ErrorMessages.UserHasMissingRoles
-import code.api.v6_0_0.APIMethods600.Implementations6_0_0
+import code.api.v6_0_0.Http4s600.Implementations6_0_0
 import code.entitlement.Entitlement
 import code.setup.DefaultUsers
 import com.github.dwickern.macros.NameOf.nameOf
