+++ obp-api/src/test/scala/code/api/v6_0_0/VerifyUserCredentialsTest.scala
@@ -7,7 +7,7 @@ import code.api.util.ApiRole
 import code.api.util.ApiRole.CanVerifyUserCredentials
 import code.api.util.ErrorMessages
 import code.api.util.ErrorMessages.{InvalidLoginCredentials, UserHasMissingRoles, UsernameHasBeenLocked}
-import code.api.v6_0_0.APIMethods600.Implementations6_0_0
+import code.api.v6_0_0.Http4s600.Implementations6_0_0
 import code.entitlement.Entitlement
 import code.loginattempts.LoginAttempt
 import code.model.dataAccess.AuthUser
